+++ R/alpha_table.R
@@ -1,7 +1,7 @@
 #' Table with alpha values
-#' 
+#'
 #' Returns a list of alpha cronbachs.
-#' 
+#'
 #' @param data A data Frame
 #' @param scales A list containing vectors with variable names. Each list element defines one scale. Named list elements are used as labels.
 #' @param labels Label names for scales (defaults to named list elements in 'scales').
@@ -15,24 +15,24 @@
 #' @param values Sets maximum and minimum valid values necessary to calculate item difficulty
 #' @export
 
-alpha_table <- function(data, scales, labels = NULL, round = 2, CI = TRUE, 
-                        conf_level = 0.95, check_key = TRUE, omega = FALSE, 
-                        keys = NULL, RMSEA = FALSE, difficulty = FALSE, 
+alpha_table <- function(data, scales, labels = NULL, round = 2, CI = TRUE,
+                        conf_level = 0.95, check_key = TRUE, omega = FALSE,
+                        keys = NULL, RMSEA = FALSE, difficulty = FALSE,
                         values = NULL) {
-  
+
   if (difficulty && is.null(values)) {
     stop("Can not calculate item difficulty without min and max scale values.")
   }
   if (is.null(labels)) labels <- labels(scales)
   df <- data.frame(Scale = labels)
-  
+
   for (i in 1:length(scales)) {
     if (!is.null(keys)) key <- keys[[i]] else key <- NULL
     if (!is.null(values)) {
-      min <- values[[i]][1] 
-      max <- values[[i]][2] 
+      min <- values[[i]][1]
+      max <- values[[i]][2]
     }
-    
+
     a <- invisible(psych::alpha(data[, scales[[i]]], check.key = check_key, keys = key))
     if (omega) {
       o <- invisible(psych::omega(data[, scales[[i]]], nfactors = 1, keys = key))
@@ -40,33 +40,33 @@ alpha_table <- function(data, scales, labels = NULL, round = 2, CI = TRUE,
     f <- invisible(psych::fa(data[, scales[[i]]]))
     alpha <- a$total$raw_alpha
     df$"n"[i] <- min(a$item.stats$n, na.rm = TRUE)
-    
+
     df$"n items"[i] <- a$nvar
-    
+
     if (!CI) df$Alpha[i] <- substring(round(alpha, round), 2)
-    
+
     if (CI) {
       a.CI <- .alpha_CI(alpha, nrow(data[, scales[[i]]]), length(scales[[i]]), conf_level)
       df$Alpha[i] <- paste0(substring(round(alpha, round), 2), " (", substring(round(a.CI[1], round), 2), "-", substring(round(a.CI[2], round), 2), ")")
     }
-    
+
     alpha.std <- a$total$std.alpha
     if (!CI) {
       df$"Std.Alpha"[i] <- substring(round(alpha.std, round), 2)
     }
-    
+
     if (CI) {
       a.std.CI <- .alpha_CI(alpha.std, nrow(data[, scales[[i]]]), length(scales[[i]]), conf_level)
       df$"Std.Alpha"[i] <- paste0(
-          substring(round(alpha.std, round), 2), 
-          " (", 
-          substring(round(a.std.CI[1], round), 2), 
-          "-", 
-          substring(round(a.std.CI[2], round), 2), 
+          substring(round(alpha.std, round), 2),
+          " (",
+          substring(round(a.std.CI[1], round), 2),
+          "-",
+          substring(round(a.std.CI[2], round), 2),
           ")"
       )
     }
-    
+
     dmin <- round(min(a$item.stats$r.drop), round)
     dmax <- round(max(a$item.stats$r.drop), round)
     mmin <- round(min(a$item.stats$mean), round)
@@ -88,13 +88,13 @@ alpha_table <- function(data, scales, labels = NULL, round = 2, CI = TRUE,
     df$"|Loading|"[i] <- paste0(substring(lmin, 2), " - ", substring(lmax, 2))
     if (RMSEA) df$"RMSEA"[i] <- substring(round(f$RMSEA[1], 3), 2)
   }
-  
+
   if (CI) {
     names(df)[which(names(df) == "Alpha")] <- paste0("Alpha(CI", conf_level * 100, "%)")
     names(df)[which(names(df) == "Std.Alpha")] <- paste0("Std.Alpha(CI", conf_level * 100, "%)")
   }
-  
-  return(df)
+
+  df
 }
 
 .alpha_CI <- function(alpha, n, items, ci) {